consolidate to 'spatial' sub folder
