+++ spiketools/plts/spatial.py
